Built a very simple web browser using windows forms
Following a tutorial, I learned how to build a simple web browser using
Windows Forms. This might be something I would like to continue building
up, but it is not something I feel like I need to use. However, if I
need to include a web browser in a piece of software, I potentially
could, like in Windows Phone or something. It's very interesting.